+++ src/plots/donut/component/ring-statistic.ts
@@ -1,13 +1,20 @@
 import { View } from '../../../dependents';
 import StatisticHtml, { IStatisticHtml } from './statistic';
-import * as statisticTemplate from './statistic-template';
-import { debounce, each, isString, isObject, isFunction, keys } from '@antv/util';
-import { LooseMap } from '../../../interface/types';
+import { getTemplate } from './statistic-template';
+import { debounce, each } from '@antv/util';
 import Ring, { DonutViewConfig } from '../layer';
+import { LooseMap } from '../../../interface/types';
 
 interface IRingStatistic extends IStatisticHtml {
   view: View;
   plot: any;
+}
+
+interface StatisticData {
+  name: string;
+  value: string;
+  itemData?: LooseMap;
+  color?: string;
 }
 
 export default class RingStatistic extends StatisticHtml {
@@ -29,7 +36,7 @@ export default class RingStatistic extends StatisticHtml {
     this.view.on(
       `interval:${triggerOnEvent}`,
       debounce((e) => {
-        const displayData = this.parseStatisticData(e.data.data);
+        const displayData = this.parseStatisticData('item', e.data.data, e.data.color);
         const htmlString = this.getStatisticHtmlString(displayData);
         this.updateHtml(htmlString);
       }, 150)
@@ -51,12 +58,12 @@ export default class RingStatistic extends StatisticHtml {
     } else {
       /** 用户没有指定文本内容时，默认显示总计 */
       const data = this.getTotalValue();
-      displayData = this.parseStatisticData(data);
+      displayData = this.parseStatisticData('total', data);
     }
     /** 中心文本显示 */
     let htmlString;
     if (this.options.htmlContent) {
-      htmlString = this.options.htmlContent(displayData);
+      htmlString = this.options.htmlContent(displayData, this.getStatisticSize());
     } else {
       htmlString = this.getStatisticTemplate(displayData);
     }
@@ -87,33 +94,29 @@ export default class RingStatistic extends StatisticHtml {
     return data;
   }
 
-  private parseStatisticData(data) {
+  private parseStatisticData(type: string, data, color?: string) {
     const plot = this.plot;
     const { angleField, colorField } = plot.options;
     const angleScale = plot.getScaleByField(angleField);
     const colorScale = plot.getScaleByField(colorField);
 
-    if (colorField) {
-      return {
-        name: colorScale.getText(data[colorField]),
-        value: angleScale.getText(data[angleField]),
-      };
+    const statisticData: StatisticData = {
+      name: colorScale ? colorScale.getText(data[colorField]) : null,
+      value: angleScale.getText(data[angleField]),
+    };
+
+    if (type === 'item') {
+      // 每一个扇形区域的数据
+      statisticData.itemData = data;
+      statisticData.color = color;
     }
 
-    return angleScale.getText(data[angleField]);
+    return statisticData;
   }
 
   private getStatisticTemplate(data) {
     const size = this.getStatisticSize();
-    let htmlString;
-    /** 如果文本内容为string或单条数据 */
-    if (isString(data)) {
-      htmlString = statisticTemplate.getSingleDataTemplate(data, this.statisticClass, size);
-    } else if (isObject(data) && keys(data).length === 2) {
-      /** 如果文本内容为两条数据 */
-      const content = data as LooseMap;
-      htmlString = statisticTemplate.getTwoDataTemplate(content.name, content.value, this.statisticClass, size);
-    }
+    const htmlString = getTemplate(data.name, data.value, this.statisticClass, size);
     /** 更为复杂的文本要求用户自行制定html模板 */
     return htmlString;
   }
@@ -127,15 +130,14 @@ export default class RingStatistic extends StatisticHtml {
   }
 
   private getStatisticHtmlString(data): string {
-    const triggerOnConfig = this.options.triggerOn;
+    const htmlContent = this.options.htmlContent;
     let htmlString: string;
-    if (isString(triggerOnConfig)) {
+    if (htmlContent) {
+      htmlString = htmlContent(data, this.getStatisticSize());
+    } else {
       htmlString = this.getStatisticTemplate(data);
     }
-    if (isFunction(triggerOnConfig)) {
-      htmlString = triggerOnConfig(data);
-      htmlString = `<div class="ring-guide-html ${this.statisticClass}">${htmlString}</div>`;
-    }
+
     return htmlString;
   }
 }